Share push request guard across API handlers
