+++ _api/push/unregister.ts
@@ -1,10 +1,5 @@
 import type { VercelRequest, VercelResponse } from "@vercel/node";
 import { validateAuth } from "../_utils/auth/index.js";
-import {
-  getEffectiveOrigin,
-  isAllowedOrigin,
-  setCorsHeaders,
-} from "../_utils/_cors.js";
 import { initLogger } from "../_utils/_logging.js";
 import { getPushMetadataLookupConcurrency } from "./_config.js";
 import {
@@ -13,6 +8,7 @@ import {
 } from "./_errors.js";
 import { getTokenOwnershipEntries, splitTokenOwnership } from "./_ownership.js";
 import { normalizeUnregisterPushPayload } from "./_request-payloads.js";
+import { handlePushPostRequestGuards } from "./_request-guard.js";
 import { createPushRedis, getMissingPushRedisEnvVars } from "./_redis.js";
 import {
   removeTokenMetadataKeys,
@@ -35,31 +31,17 @@ export const maxDuration = 15;
 export default async function handler(req: VercelRequest, res: VercelResponse) {
   const { logger } = initLogger();
   const startTime = Date.now();
-  const origin = getEffectiveOrigin(req);
 
-  logger.request(req.method || "POST", req.url || "/api/push/unregister");
-
-  if (req.method === "OPTIONS") {
-    if (!isAllowedOrigin(origin)) {
-      logger.response(403, Date.now() - startTime);
-      return res.status(403).json({ error: "Unauthorized" });
-    }
-    setCorsHeaders(res, origin, { methods: ["POST", "OPTIONS"] });
-    logger.response(204, Date.now() - startTime);
-    return res.status(204).end();
-  }
-
-  if (!isAllowedOrigin(origin)) {
-    logger.response(403, Date.now() - startTime);
-    return res.status(403).json({ error: "Unauthorized" });
-  }
-
-  setCorsHeaders(res, origin, { methods: ["POST", "OPTIONS"] });
-
-  if (req.method !== "POST") {
-    res.setHeader("Allow", "POST, OPTIONS");
-    logger.response(405, Date.now() - startTime);
-    return res.status(405).json({ error: "Method not allowed" });
+  if (
+    handlePushPostRequestGuards(
+      req,
+      res,
+      logger,
+      startTime,
+      "/api/push/unregister"
+    )
+  ) {
+    return;
   }
 
   try {
